pyproject.toml
[tool.poetry]
name = "python-flask-template"
version = "0.1.0"
description = ""
authors = ["Shonte Amato-Grill <shonte.amatogrill@gmail.com>"]
readme = "README.md"
packages = [{include = "surgicai_api"}]

[tool.poetry.dependencies]
python = "^3.10"
gunicorn = "^20.1.0"
flask = "^2.2.2"
environs = "^14.2.0"
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.0"
jinja2 = "^3.1.6"
pyjwt = "^2.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
black = "^22.12.0"
pylint = "^2.15.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
